Fixed KeeperController and OwnerController.
