+++ Views/owner-add.php
@@ -7,7 +7,7 @@
      <section id="listado" class="mb-5">
           <div class="container">
                <h2 class="mb-4">New Owner</h2>
-               <form action="<?php echo FRONT_ROOT ?>User/newOwner" method="post" class="bg-light-alpha p-5">
+               <form action="<?php echo FRONT_ROOT ?>Owner/newOwner" method="post" class="bg-light-alpha p-5">
                     <div class="row">                         
                       <!--   <div class="col-lg-4">
                               <div class="form-group">
